refactor(tutorial-service): improve error handling in StateManager and NavigationManager, and standardize import paths
