+++ src/ui/tutorial/controller/external.ts
@@ -1,6 +1,6 @@
 import { Tutorial } from '@domain/models/Tutorial';
 import { IUserInteraction } from '@domain/ports/IUserInteraction';
-import { TutorialService } from '@domain/services/tutorial-service/TutorialService';
+import { TutorialService } from '@domain/services/tutorial-service';
 import * as vscode from 'vscode';
 
 export type Args = {
